Crud terminado y diseño de email

+++ FrontEnd2/src/app/shared/components/user/user.component.ts
@@ -1,9 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import {UserAddComponent} from '../user/user-add/user-add.component'
+import { UserAddComponent } from '../user/user-add/user-add.component'
+import { UserEditComponent } from '../user/user-edit/user-edit.component'
 import { UserI } from '../../models/user.interface'
 import { UserServiceService } from '../../service/SUser/user-service.service';
 import { MatDialog } from '@angular/material/dialog';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-user',
@@ -14,46 +16,84 @@ export class UserComponent implements OnInit {
 
   users: UserI[];
 
-  constructor(private service:UserServiceService, private router: Router, 
-    public dialog:MatDialog) { }
+  constructor(private service: UserServiceService, private router: Router,
+    public dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.getData();
   }
 
-  getData(){
+  getData() {
     this.service.getUsers().subscribe((result: any) => {
       this.users = result;
-      console.log( this.users);
+      console.log(this.users);
     })
   }
 
-  deleteUser(id:number){
-    this.service.deleteUser(id).subscribe((res:any)=>{
-    this.router.navigate(['/user']);
-   })
+  deleteUser(id: number) {
+
+    Swal.fire({
+      title: `Esta seguro de eliminar el usuario con la cedula ${id}?`,
+      text: "Esta acción es irreversible!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          position: 'center',
+          icon: 'success',
+          title: 'Eliminado exitosamente',
+          showConfirmButton: false,
+          timer: 1500
+        })
+      }
+      this.service.deleteUser(id).subscribe((res: any) => {
+        this.ngOnInit();
+      })
+    })
   }
 
-  editUser(id:number){
-
+  editUser(user:any) {
+    console.log(user);
+  this. openEdit(user);
   }
 
-  addUser(){
+  addUser() {
     this.openAdd();
   }
 
-  openAdd():void{
-    console.log('HOLI');
-    const dialogRef = this.dialog.open(UserAddComponent,{
-      height:'450px',
-      width:'900px',
+  openAdd(): void {
+    const dialogRef = this.dialog.open(UserAddComponent, {
+      height: '450px',
+      width: '900px',
     });
 
-    dialogRef.afterClosed().subscribe(result =>{
+    dialogRef.afterClosed().subscribe(result => {
 
       console.log(`Dialog result ${result}`);
+      this.ngOnInit();
 
     })
+  }
+
+  openEdit(user: any): void {
+    const dialogRef = this.dialog.open(UserEditComponent, {
+      height: '450px',
+      width: '900px',
+      data: {
+        user:user
+      }
+    });
+    dialogRef.afterClosed().subscribe(result => {
+
+      console.log(`Dialog result ${result}`);
+      this.ngOnInit();
+
+    })
+
   }
 
 }
